requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.8.0
fsspec==2025.3.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
groq==0.19.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.7.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
openai==2.8.1
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==5.29.3
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
twilio==9.8.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.38.0
virtualenv==20.34.0
Werkzeug==3.1.3
whitenoise==6.11.0
wrapt==1.17.2
yarl==1.22.0
